+++ apps/e2e-tests/playwright.config.ts
@@ -22,6 +22,7 @@ console.log("Running tests for: ", baseURL);
 
 const config: PlaywrightTestConfig = {
   testDir: "./tests",
+  outputDir: "./reports",
   /* Maximum time one test can run for. */
   timeout: 90 * 1000,
   expect: {