+++ examples/modal-teleport/components/AppModal.vue
@@ -9,12 +9,12 @@ const { isOpen, close } = props.controller;
 <template>
   <Teleport to="body">
     <transition
-      enter-active-class="transition ease-out duration-200 transform"
-      enter-from-class="opacity-0"
-      enter-to-class="opacity-100"
-      leave-active-class="transition ease-in duration-200 transform"
-      leave-from-class="opacity-100"
-      leave-to-class="opacity-0"
+      enterActiveClass="transition ease-out duration-200 transform"
+      enterFromClass="opacity-0"
+      enterToClass="opacity-100"
+      leaveActiveClass="transition ease-in duration-200 transform"
+      leaveFromClass="opacity-100"
+      leaveToClass="opacity-0"
     >
       <div
         class="fixed z-10 inset-0 overflow-y-auto bg-black bg-opacity-50"
@@ -25,12 +25,12 @@ const { isOpen, close } = props.controller;
           class="flex items-start justify-center min-h-screen pt-24 text-center"
         >
           <transition
-            enter-active-class="transition ease-out duration-300 transform "
-            enter-from-class="opacity-0 translate-y-10 scale-95"
-            enter-to-class="opacity-100 translate-y-0 scale-100"
-            leave-active-class="ease-in duration-200"
-            leave-from-class="opacity-100 translate-y-0 scale-100"
-            leave-to-class="opacity-0 translate-y-10 translate-y-0 scale-95"
+            enterActiveClass="transition ease-out duration-300 transform "
+            enterFromClass="opacity-0 translate-y-10 scale-95"
+            enterToClass="opacity-100 translate-y-0 scale-100"
+            leaveActiveClass="ease-in duration-200"
+            leaveFromClass="opacity-100 translate-y-0 scale-100"
+            leaveToClass="opacity-0 translate-y-10 translate-y-0 scale-95"
           >
             <div
               class="bg-white rounded-lg text-left overflow-hidden shadow-xl p-8 w-1/2"