+++ packages/cms-base/components/public/cms/element/CmsElementProductDescriptionReviews.vue
@@ -5,29 +5,28 @@ import {
   getProductName,
 } from "@shopware-pwa/helpers-next";
 import SwProductReviews from "../../../SwProductReviews.vue";
-import deepMerge from '../../../../helpers/deepMerge'
+import deepMerge from "../../../../helpers/deepMerge";
 import getTranslations from "../../../../helpers/getTranslations";
 
-const props = 
-  defineProps<{
-     content: CmsElementProductDescriptionReviews;
-  }>();
+const props = defineProps<{
+  content: CmsElementProductDescriptionReviews;
+}>();
 
 type Translations = {
   products: {
     description: string;
     reviews: string;
-  }
-}
+  };
+};
 
 let translations: Translations = {
   products: {
     description: "Description",
-    reviews: "Reviews"
-  }
-}
-const globalTranslations = getTranslations()
-translations = deepMerge(translations, globalTranslations) as Translations
+    reviews: "Reviews",
+  },
+};
+const globalTranslations = getTranslations();
+translations = deepMerge(translations, globalTranslations) as Translations;
 
 const currentTab = ref<number>(1);
 const { product } = useProduct(props.content.data?.product);
@@ -62,7 +61,8 @@ const reviews = computed(() => props.content.data.reviews?.elements);
             ]"
             @click="() => toggleTabs(1)"
           >
-            <i class="fas fa-space-shuttle text-base mr-1" /> {{translations.products.description}}
+            <i class="fas fa-space-shuttle text-base mr-1" />
+            {{ translations.products.description }}
           </a>
         </li>
         <li class="mr-2 text-center">
@@ -75,7 +75,8 @@ const reviews = computed(() => props.content.data.reviews?.elements);
             ]"
             @click="() => toggleTabs(2)"
           >
-            <i class="fas fa-cog text-base mr-1" /> {{translations.products.reviews}}
+            <i class="fas fa-cog text-base mr-1" />
+            {{ translations.products.reviews }}
           </a>
         </li>
       </ul>