+++ packages/types/shopware-6-client/models/checkout/shipping/ShippingMethod.d.ts
@@ -7,6 +7,7 @@ import { Media } from "../../content/media/Media";
 import { Tag } from "../../system/tag/Tag";
 import { ShippingMethodTranslation } from "./ShippingMethodTranslation";
 import { CustomFields } from "../../common/CustomField";
+import { Tax } from "../../system/tax/Tax";
 
 /**
  * @public
@@ -16,14 +17,14 @@ export type ShippingMethod = {
   name: string | null;
   active: boolean;
   position?: number;
-  customFields?: CustomFields;
+  customFields?: CustomFields | null;
   mediaId: string | null;
   deliveryTimeId: string;
   taxType: string;
   description: string | null;
   trackingUrl: string | null;
-  createdAt: Date;
-  updatedAt: Date;
+  createdAt: string | Date;
+  updatedAt: Date | string | null;
   translated: ShippingMethodTranslation | null;
   deliveryTime: DeliveryTime | null;
   availabilityRule: Rule | null;
@@ -31,8 +32,10 @@ export type ShippingMethod = {
   media: Media | null;
   tags: Tag[] | null;
   translations: ShippingMethodTranslation[] | null;
-  orderDeliveries: OrderDelivery[] | null;
-  salesChannelDefaultAssignments: SalesChannel[] | null;
-  salesChannels: SalesChannel[] | null;
-  availabilityRuleId: string;
+  orderDeliveries?: OrderDelivery[] | null;
+  salesChannelDefaultAssignments?: SalesChannel[] | null;
+  salesChannels?: SalesChannel[] | null;
+  availabilityRuleId?: string;
+  tax: Tax | null;
+  apiAlias: "shipping_method";
 };