+++ packages/types/shopware-6-client/models/checkout/shipping/ShippingMethodTranslation.d.ts
@@ -5,9 +5,10 @@ import { CustomFields } from "../../common/CustomField";
  * @public
  */
 export type ShippingMethodTranslation = {
-  shippingMethodId: string;
+  shippingMethodId?: string | null;
   name: string | null;
   description: string | null;
-  shippingMethod: ShippingMethod | null;
+  shippingMethod?: ShippingMethod | null;
   customFields: CustomFields;
+  trackingUrl: string | null;
 };