+++ templates/vue-demo-store/components/account/AccountAddressForm.vue
@@ -68,7 +68,7 @@ useFocus(firstNameInputElement, { initialValue: true });
                 for="country"
                 class="block mb-2 text-sm font-medium text-gray-500"
               >
-                {{ $t('form.salutation') }}
+                {{ $t("form.salutation") }}
               </label>
               <select
                 id="salutation"
@@ -94,7 +94,7 @@ useFocus(firstNameInputElement, { initialValue: true });
                 for="first-name"
                 class="block mb-2 text-sm font-medium text-gray-500"
               >
-                {{ $t('form.firstName') }}
+                {{ $t("form.firstName") }}
               </label>
               <input
                 id="first-name"
@@ -113,7 +113,7 @@ useFocus(firstNameInputElement, { initialValue: true });
                 for="last-name"
                 class="block mb-2 text-sm font-medium text-gray-500"
               >
-               {{ $t('form.lastName') }}
+                {{ $t("form.lastName") }}
               </label>
               <input
                 id="last-name"
@@ -130,7 +130,7 @@ useFocus(firstNameInputElement, { initialValue: true });
                 for="country"
                 class="block mb-2 text-sm font-medium text-gray-500"
               >
-                  {{ $t('form.country') }}
+                {{ $t("form.country") }}
               </label>
               <select
                 id="country"
@@ -157,7 +157,7 @@ useFocus(firstNameInputElement, { initialValue: true });
                 for="street-address"
                 class="block mb-2 text-sm font-medium text-gray-500"
               >
-                 {{ $t('form.streetAddress') }}
+                {{ $t("form.streetAddress") }}
               </label>
               <input
                 id="street-address"
@@ -176,7 +176,7 @@ useFocus(firstNameInputElement, { initialValue: true });
                 for="city"
                 class="block mb-2 text-sm font-medium text-gray-500"
               >
-                 {{ $t('form.city') }}
+                {{ $t("form.city") }}
               </label>
               <input
                 id="city"
@@ -194,7 +194,7 @@ useFocus(firstNameInputElement, { initialValue: true });
                 for="postal-code"
                 class="block mb-2 text-sm font-medium text-gray-500"
               >
-                {{ $t('form.postalCode') }}
+                {{ $t("form.postalCode") }}
               </label>
               <input
                 id="postal-code"
@@ -216,7 +216,7 @@ useFocus(firstNameInputElement, { initialValue: true });
             data-testid="account-address-form-submit-button"
             @click.stop.prevent="invokeSave"
           >
-            {{ $t('form.save') }}
+            {{ $t("form.save") }}
           </button>
         </div>
       </form>