+++ templates/vue-demo-store/nuxt.config.ts
@@ -12,23 +12,15 @@ export default defineNuxtConfig({
       },
     },
   },
-  alias: {
-    /**
-     * TODO: Temp fix until new VueUse published:
-     * - https://github.com/vueuse/vueuse/pull/2449
-     * - https://github.com/vueuse/vueuse/actions/workflows/publish.yml
-     */
-    useMeta: "~/composables/useMeta",
-  },
   /**
    * Commented because of the StackBlitz error
    * Issue: https://github.com/shopware/frontends/issues/88
    */
-  // typescript: {
-  //   typeCheck: true,
-  //   strict: true,
-  // },
   debug: true,
+  typescript: {
+    // typeCheck: true,
+    strict: true,
+  },
   modules: [
     "@vueuse/nuxt",
     "@unocss/nuxt",