+++ templates/vue-demo-store/pages/checkout/index.vue
@@ -7,6 +7,7 @@ export default {
 import { useVuelidate } from "@vuelidate/core";
 import { required, email, minLength, requiredIf } from "@vuelidate/validators";
 import { ClientApiError, ShopwareError } from "@shopware-pwa/types";
+import { getShippingMethodDeliveryTime } from "@shopware-pwa/helpers-next";
 
 definePageMeta({
   layout: "checkout",
@@ -624,6 +625,12 @@ const addAddressModalController = useModal();
                   <div>
                     {{ singleShippingMethod.translated?.name }}
                     <span
+                      v-if="getShippingMethodDeliveryTime(singleShippingMethod)"
+                      >({{
+                        getShippingMethodDeliveryTime(singleShippingMethod)
+                      }})</span
+                    >
+                    <span
                       v-if="singleShippingMethod.translated?.description"
                       class="italic text-sm text-gray-500 block"
                     >
@@ -678,9 +685,27 @@ const addAddressModalController = useModal();
               <label
                 :for="singlePaymentMethod.id"
                 :class="{ 'animate-pulse': isLoading[singlePaymentMethod.id] }"
-                class="ml-2 block text-sm font-medium text-gray-700"
+                class="ml-2 block text-sm font-medium text-gray-700 w-full"
               >
-                {{ singlePaymentMethod.translated?.name }}
+                <div class="flex justify-between">
+                  <div>
+                    <span>
+                      {{ singlePaymentMethod.translated?.name }}
+                    </span>
+                    <span
+                      v-if="singlePaymentMethod.translated?.description"
+                      class="italic text-sm text-gray-500 block"
+                    >
+                      {{ singlePaymentMethod.translated.description }}</span
+                    >
+                  </div>
+                  <div v-if="singlePaymentMethod.media?.url">
+                    <img
+                      :src="singlePaymentMethod.media.url"
+                      alt="payment-image"
+                    />
+                  </div>
+                </div>
               </label>
             </div>
           </fieldset>